# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API for controlling the observability collector
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Observability Collector API
  version: "1.0"
paths:
...
  /buffers/{name}:
    get:
      consumes:
      - application/json
      description: Get information about a specific buffer
      parameters:
      - description: Buffer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get buffer by name
      tags:
      - buffers
  /buffers/{name}/flush:
    post:
      consumes:
      - application/json
      description: Flush a specific buffer
      parameters:
      - description: Buffer name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Flush a buffer
      tags:
      - buffers
  /config:
    get:
      consumes:
      - application/json
      description: Get the current collector configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get configuration
      tags:
      - config
    put:
      consumes:
      - application/json
      description: Update the collector configuration
      parameters:
      - description: New configuration
        in: body
        name: config
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update configuration
      tags:
      - config
...
  /pipelines:
    get:
      consumes:
      - application/json
      description: Get information about all data pipelines
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all pipelines
      tags:
      - pipelines
    post:
      consumes:
      - application/json
      description: Create a new data pipeline
      parameters:
      - description: Pipeline configuration
        in: body
        name: pipeline
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a pipeline
      tags:
      - pipelines
  /pipelines/{type}:
    delete:
      consumes:
      - application/json
      description: Delete a data pipeline by telemetry type
      parameters:
      - description: Pipeline type (logs, metrics, traces)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a pipeline
      tags:
      - pipelines
    get:
      consumes:
      - application/json
      description: Get information about a specific pipeline by telemetry type
      parameters:
      - description: Pipeline type (logs, metrics, traces)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get pipeline by type
      tags:
      - pipelines
...
  /plugins/{type}:
    get:
      consumes:
      - application/json
      description: Get information about plugins of a specific type
      parameters:
      - description: Plugin type (input, processor, output)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get plugins by type
      tags:
      - plugins
  /plugins/{type}/{name}:
    get:
      consumes:
      - application/json
      description: Get information about a specific plugin
      parameters:
      - description: Plugin type (input, processor, output)
        in: path
        name: type
        required: true
        type: string
      - description: Plugin name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get plugin by name
      tags:
      - plugins
  /plugins/{type}/{name}/restart:
    post:
      consumes:
      - application/json
      description: Restart a specific plugin
      parameters:
      - description: Plugin type (input, processor, output)
        in: path
        name: type
        required: true
        type: string
      - description: Plugin name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Restart a plugin
      tags:
      - plugins
  /plugins/{type}/{name}/start:
    post:
      consumes:
      - application/json
      description: Start a specific plugin
      parameters:
      - description: Plugin type (input, processor, output)
        in: path
        name: type
        required: true
        type: string
      - description: Plugin name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a plugin
      tags:
      - plugins
  /plugins/{type}/{name}/stop:
    post:
      consumes:
      - application/json
      description: Stop a specific plugin
      parameters:
      - description: Plugin type (input, processor, output)
        in: path
        name: type
        required: true
        type: string
      - description: Plugin name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stop a plugin
      tags:
      - plugins
  /restart:
    post:
      consumes:
      - application/json
      description: Restart the observability collector
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Restart collector
      tags:
      - control
  /start:
    post:
      consumes:
      - application/json
      description: Start the observability collector
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start collector
      tags:
      - control
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"